implement global exception handling

+++ src/main/java/com/Skill/Marketplace/SM/Services/SkillService.java
@@ -3,6 +3,8 @@ import com.Skill.Marketplace.SM.DTO.skillDTO.CreateSkillDTO;
 import com.Skill.Marketplace.SM.DTO.skillDTO.UpdateSkillDTO;
 import com.Skill.Marketplace.SM.Entities.Category;
 import com.Skill.Marketplace.SM.Entities.Skill;
+import com.Skill.Marketplace.SM.Exception.BadRequestException;
+import com.Skill.Marketplace.SM.Exception.ResourceNotFoundException;
 import com.Skill.Marketplace.SM.Repo.CategoryRepo;
 import com.Skill.Marketplace.SM.Repo.SkillsRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,8 +23,12 @@ public class SkillService {
 
     public Skill create(CreateSkillDTO dto){
 
+        if(dto.getSkillName()==null || dto.getSkillName().isEmpty()){
+            throw new BadRequestException("Skill name is required");
+        }
+
         Category category = categoryRepo.findById(dto.getCategoryId())
-                .orElseThrow(()-> new RuntimeException("No Category found"));
+                .orElseThrow(()-> new ResourceNotFoundException("No Category found"));
 
         Skill skill = new Skill();
         skill.setSkillName(dto.getSkillName());
@@ -34,7 +40,7 @@ public class SkillService {
 
     public Skill getById(Long id){
         return skillsRepo.findById(id)
-                .orElseThrow(()->new RuntimeException("No skills found"));
+                .orElseThrow(()->new ResourceNotFoundException("No skills found"));
     }
 
     public List<Skill> getAll(){
@@ -43,13 +49,17 @@ public class SkillService {
 
     public Skill update(Long id , UpdateSkillDTO dto){
         Skill skill = skillsRepo.findById(id)
-                .orElseThrow(()-> new RuntimeException("No skills found"));
+                .orElseThrow(()-> new ResourceNotFoundException("No skills found"));
+
+        if(dto.getSkillName()==null || dto.getSkillName().isEmpty()){
+            throw new BadRequestException("Skill name is required");
+        }
 
         skill.setSkillName(dto.getSkillName());
 
 
         Category category =  categoryRepo.findById(dto.getCategoryId())
-                .orElseThrow(()-> new RuntimeException("No category found"));
+                .orElseThrow(()-> new ResourceNotFoundException("No category found"));
         skill.setCategory(category);
 
         return skillsRepo.save(skill);
@@ -57,6 +67,10 @@ public class SkillService {
     }
 
     public void delete(Long id){
+
+        if(!skillsRepo.existsById(id)){
+            throw new ResourceNotFoundException("No skills found");
+        }
         skillsRepo.deleteById(id);
     }
 }
