+++ src/main/java/com/Skill/Marketplace/SM/Services/UserService.java
@@ -3,6 +3,8 @@ import com.Skill.Marketplace.SM.DTO.userDTO.CreateUserDTO;
 import com.Skill.Marketplace.SM.DTO.userDTO.UpdateUserDTO;
 import com.Skill.Marketplace.SM.Entities.UserModel;
 import com.Skill.Marketplace.SM.Entities.UserType;
+import com.Skill.Marketplace.SM.Exception.BadRequestException;
+import com.Skill.Marketplace.SM.Exception.ResourceNotFoundException;
 import com.Skill.Marketplace.SM.Repo.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -18,6 +20,14 @@ public class UserService {
     private PasswordEncoder passwordEncoder;
 
     public UserModel createNewUser(CreateUserDTO dto){
+
+        if(dto.getUsername()==null || dto.getUsername().isEmpty()){
+            throw new BadRequestException("Username is required");
+        }
+
+        if(dto.getPassword()==null || dto.getPassword().isEmpty()){
+            throw new BadRequestException("Password is required");
+        }
 
         UserModel user = new UserModel();
 
@@ -40,18 +50,18 @@ public class UserService {
 
     public UserModel getUserProfile(String username){
         return userRepo.getUserByUsername(username)
-                .orElseThrow(()-> new RuntimeException("User not found"));
+                .orElseThrow(()-> new ResourceNotFoundException("User not found"));
     }
 
     public UserModel updateUser(String username , UpdateUserDTO request){
         UserModel user = userRepo.getUserByUsername(username)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
         user.setFirstName(request.getFirstName());
         user.setLastName(request.getLastName());
 
         if(request.getUserType()==UserType.ADMIN){
-            throw new RuntimeException("Cannot update user to ADMIN role");
+            throw new BadRequestException("Cannot update user to ADMIN role");
         }
         else{
             user.setUserType(request.getUserType());
@@ -61,6 +71,10 @@ public class UserService {
     }
 
     public void deleteUser(String username){
+
+        if(!userRepo.existsByUsername(username)){
+            throw new ResourceNotFoundException("User not found");
+        }
         userRepo.deleteUserByUsername(username);
     }
 }
